requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
PyYAML==3.12
simplejson==3.10.0
six==1.10.0
Werkzeug==0.12.2
